+++ src/Farmer/Builders/Builders.KeyVault.fs
@@ -30,7 +30,9 @@ type KeyVaultConfigSettings =
       /// Specifies whether Azure Disk Encryption is permitted to retrieve secrets from the vault and unwrap keys.
       AzureDiskEncryptionAccess : FeatureFlag option
       /// Specifies whether Soft Deletion is enabled for the vault
-      SoftDelete : SoftDeletionMode option }
+      SoftDelete : SoftDeletionMode option
+      /// Specifies whether Azure role based authorization is used for data retrieval instead of any access policies on the key vault.
+      RbacAuthorization : FeatureFlag option }
 
 type NetworkAcl =
     { IpRules : string list
@@ -105,6 +107,7 @@ type KeyVaultConfig =
                   TemplateDeployment = this.Access.ResourceManagerAccess
                   DiskEncryption = this.Access.AzureDiskEncryptionAccess
                   Deployment = this.Access.VirtualMachineAccess
+                  RbacAuthorization = this.Access.RbacAuthorization
                   SoftDelete = this.Access.SoftDelete
                   CreateMode =
                     match this.Policies with
@@ -219,7 +222,13 @@ type KeyVaultBuilder() =
     member __.Yield (_:unit) =
         { Name = ResourceName.Empty
           TenantId = Subscription.TenantId
-          Access = { VirtualMachineAccess = None; ResourceManagerAccess = Some Enabled; AzureDiskEncryptionAccess = None; SoftDelete = None }
+          Access = {
+            VirtualMachineAccess = None
+            RbacAuthorization = None
+            ResourceManagerAccess = Some Enabled
+            AzureDiskEncryptionAccess = None
+            SoftDelete = None
+          }
           Sku = Standard
           NetworkAcl = { IpRules = []; VnetRules = []; Bypass = None; DefaultAction = None }
           Policies = []
@@ -273,6 +282,12 @@ type KeyVaultBuilder() =
     /// Disallows Azure Disk Encyption service access to the vault.
     [<CustomOperation "disable_disk_encryption_access">]
     member __.DisableDiskEncryptionAccess(state:KeyVaultBuilderState) = { state with Access = { state.Access with AzureDiskEncryptionAccess = Some Disabled } }
+    /// Enables Azure role based authentication for access to key vault data.
+    [<CustomOperation "enable_rbac">]
+    member __.EnableRbacAuthorization(state:KeyVaultBuilderState) = { state with Access = { state.Access with RbacAuthorization = Some Enabled } }
+    /// Disables Azure role based authentication for access to key vault data.
+    [<CustomOperation "disable_rbac">]
+    member __.DisableRbacAuthorization(state:KeyVaultBuilderState) = { state with Access = { state.Access with RbacAuthorization = Some Disabled } }
     /// Enables VM access to the vault.
     [<CustomOperation "enable_soft_delete">]
     member __.EnableSoftDeletion(state:KeyVaultBuilderState) = { state with Access = { state.Access with SoftDelete = Some SoftDeletionOnly } }